Revert "Fixed confusion about enemies in fights."
This reverts commit cce0b14c667ad178dc433b7174d9629246e027f4.

+++ Source/Fight_MovementManager.h
@@ -57,7 +57,7 @@ public:
 
 
 private:
-    double distance_metric = 0;
+    int distance_metric = 0;
     Position stutter_vector_ = Positions::Origin;
     Position attune_vector_ = Positions::Origin;
     Position cohesion_vector_ = Positions::Origin;
